End of day commit 28.4

+++ ft_strlcat.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   ft_strlcat.c                                       :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: wheino <wheino@student.hive.fi>            #+#  +:+       +#+        */
+/*   By: wheino <wheino@student.hive.fi>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2025-04-16 12:39:03 by wheino            #+#    #+#             */
-/*   Updated: 2025/04/28 12:23:23 by wheino           ###   ########.fr       */
+/*   Created: 2025/04/16 12:39:03 by wheino            #+#    #+#             */
+/*   Updated: 2025/04/28 15:20:26 by wheino           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,16 +14,16 @@
 
 size_t	ft_strlcat(char *dst, const char *src, size_t dstsize)
 {
-	int	src_len;
-	int	dst_len;
+	size_t	src_len;
+	size_t	dst_len;
 	int	i;
 
-	src_len = ft_strlen((char *)src);
-	dst_len = ft_strlen(dst);
-	if (dst_len >= (int)dstsize)
+	src_len = (size_t)ft_strlen((char *)src);
+	dst_len = (size_t)ft_strlen(dst);
+	if (dst_len >= dstsize)
 		return (dstsize + src_len);
 	i = 0;
-	while ((src[i] != '\0' && (dst_len + i < (int)dstsize - 1)))
+	while ((src[i] != '\0' && (dst_len + i < dstsize - 1)))
 	{	
 		dst[i + dst_len] = src[i];
 		i++;
